refactor(04-03): co-locate concrete strategy configs into strategy files (D-14, D-15)
- Move SMA_MACDConfig into SMA_MACD_strategy.py (re-indented to TABS, validator preserved)
- Move EmptyStrategyConfig into empty_strategy.py (re-indented to TABS)
- Subclass BaseStrategyConfig imported from itrader.config
